move tokenlists
